Utilisation de l'API C pour FMOD, pour la compatibilité Code::Blocks Windows. Utilisation de GLEW pour OpenGL Windows. Gestion du temps sous Windows avec la SDL.
Résolution d'écran converties en entiers pour éviter les problèmes de virgule flottante.
Intégration de l'éditeur dans le menu principal. Possibilité d'ajouter des niveaux.
Annulation/rétablissement dans l'éditeur, support du copier/coller !
Lors de la création d'une entité, les dernières informations sont utilisées.

Correction de divers problèmes lors de la mort des personnages.
Citation des auteurs dans les crédits.
Correction d'un bug avec la conversion UTF-16 vers UTF-8 (Session.cpp:105).

+++ Projet2MIC/src/Texte.h
@@ -15,7 +15,7 @@
 #include "Unichar.h"
 #include "Session.h"
 
-typedef unsigned int taillePolice_t;
+typedef float taillePolice_t;
 enum police_t {premierePolice, POLICE_NORMALE = premierePolice, POLICE_DECO, POLICE_GRANDE, nbPolices};
 police_t &operator++(police_t &);
